+++ tests/unit/test_duckdb.py
@@ -1,7 +1,7 @@
 import os
 
 import pytest
-
+import pandas as pd
 from viadot.sources.duckdb import DuckDB
 
 TABLE = "test_table"
@@ -74,4 +74,19 @@ def test__check_if_table_exists(duckdb, TEST_PARQUET_FILE_PATH):
         schema=SCHEMA, table=TABLE, path=TEST_PARQUET_FILE_PATH
     )
     assert duckdb._check_if_table_exists(TABLE, schema=SCHEMA)
+    duckdb.drop_table(TABLE, schema=SCHEMA)
+
+
+def test_run_query_with_comments(duckdb, TEST_PARQUET_FILE_PATH):
+    duckdb.create_table_from_parquet(
+        schema=SCHEMA, table=TABLE, path=TEST_PARQUET_FILE_PATH
+    )
+    output = duckdb.run(
+        query=f""" 
+        --test 
+    SELECT * FROM {SCHEMA}.{TABLE}
+    """,
+        fetch_type="dataframe",
+    )
+    assert isinstance(output, pd.DataFrame)
     duckdb.drop_table(TABLE, schema=SCHEMA)